Implement ToggleBookmark page

+++ lib/actions/ToogleBookmark.action.ts
@@ -18,31 +18,38 @@ export default async function ToogleBookmarkAction(params: {
 	details?: object | null;
 }> {
 	await dbConnect();
-	const auth_session = await auth();
-	const userEmail = auth_session?.user?.email;
+
 	try {
+		const auth_session = await auth();
+		const userEmail = auth_session?.user?.email;
+
 		const validatedData = validateBody(params, BookmarkSchema);
 		const { questionId } = validatedData.data;
-		const userId = await User.findOne({ email: userEmail }).select("_id");
-		const question = await Question.findById(questionId);
+
+		const user = await User.findOne({ email: userEmail }).select("_id");
+		if (!user) throw new Error("Unauthorized");
+
+		const question = await Question.findById(questionId).select("_id");
 		if (!question) throw new Error("Question not found");
+
 		const collection = await Collection.findOne({
 			question: questionId,
-			authro: userId,
+			author: user._id,
 		});
+
 		if (collection) {
 			await Collection.findByIdAndDelete(collection._id);
 			return {
 				success: true,
 				data: { saved: false },
 			};
-		} else {
-			await Collection.create({ question: questionId, author: userId });
-			return {
-				success: true,
-				data: { saved: true },
-			};
 		}
+		await Collection.create({ question: questionId, author: user._id });
+
+		return {
+			success: true,
+			data: { saved: true },
+		};
 	} catch (error) {
 		return handleActionErrorResponse(error);
 	}
